requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
email-reply-parser==0.5.12
fastapi==0.115.13
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ollama==0.5.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
